Prevents invalid sentences from causing crash
Improves the current implementation of the parse() method to scan
the entire sentence and count the fields before attempting to read
the fields.  This prevents accidental invalid pointer accesses when
parsing invalid sentences with valid checksums.

The new parse implementation also has a smaller compiled size.

Includes a sketch in the "extras" directory (ignored by the Arduino
IDE according to the library spec) that was used for testing the
parsing of all sentences known to the parser.